Refactor: Move LinkMaker to common
